Fix redirects broken by Terms and Conditions redirect (#8036)
* Fix redirects broken by Terms and Conditions redirect

* Fix redirects broken by Terms and Conditions redirect - improvements
